Create zone controller object
This object has the following responsibilities:
* It increments the stage timer and kills the player in the event of a "time over".
* It culls all instances periodically (currently every 3 steps).
* It stores information about the name and act of the current zone, for use with the title card and score card (objects which will be added sometime in the future).
* It resets the player's ring count at the end of the stage (it may also reset other values, but this will be also be determined in the future).

It was previously highlighted in the game initialization script that I had intended to move the queueing of the Mad Gear Zone background music track to the zone controller object, specifically its Room Start event; however, this proved not to be possible as the music controller object is not yet created when Room Start events are executed (which was highlighted in an author note left in the window controller object's Create event) so the music track ultimately never plays. For now, its queueing remains in the script, and it will be decided at a later time on whether it should remain like this, or a dedicated startup room should be added in which global controllers are created.

Furthermore, the functionality of the player's mask rotation method has been slightly improved: upon landing, the `rotation_lock_time` static variable is set to 0 to allow the player's virtual mask to rotate on the next frame they are grounded, if applicable.