Create HUD object
The object is created in the zone controller object's Room Start event, and is assigned to a "Display" instance layer that has been added to the test room. Objects which render UI elements or overlay themselves on top of the screen (such as a title card and room fade transition) will be assigned to this layer when they are added in the future.

The fonts used to display the player's score/time/rings and lives have also been added and initialized, and the game GUI size has been defined by the window controller object in its Create event, to allow the HUD's scale to remain constant when scaling the window.

The player's debug overlay is now positioned to the right, and its text has been shrunk due to the GUI size changing.