Add tile collisions (pt. 2)
Support for four tile layers has been added; these are stored in an array initialized by the player object, and match the properties of solid zone objects:
- Main (not semisolid or bound to a specific collision layer)
- Semisolid
- Layer 0
- Layer 1

This array is refreshed prior to solid collision detection in the player's movement step, and, just like with solid zone objects, if the player cannot interact with a certain tile layer, it is removed from the array. In part of this, to make the removal of layers 0 & 1 smoother, the player's `collision_layer` variable is now limited to 0 & 1, instead of 1 & 2. Furthermore, since solid zone objects will no longer be used to make up the terrain for stages, their `collision_layer` variable has been removed, and the "Terrain" layer from the test stage has been deleted.

Functions which checked for collision with solid zone objects have been updated to check for collision against these tile layers, and in some cases have been combined into a single loop (notably angle & cliff detection).

It was no longer appropriate for the player to use a rectangle for detecting floors and ceilings, as they would clip into them by an additional pixel, and suffer from angle detection issues. This is in part due to the `collision_rectangle` function needing to overlap the given tilemap element or instance's mask by a minimum of 0.5 pixels for a collision to be registered. `collision_line` is not bounded by this threshold, and thus, is now used for the aforementioned scenario. This ultimately leaves the `player_body_in_object` function unused, however, it will remain for zone object reactions until further notice.

In the wake of these changes, the naming of the player's collision functions was brought into question. To make the rectangle to line transition smoother for floor and ceiling detection, the `player_arms_in_object` function was rewritten accordingly, now accepting a variable diameter for the line and vertical offset, but this has caused it to lose accuracy in its naming. The `player_leg_in_object` function has faced a similar issue, whereby passing values other than the player's x-radius to the `xoff` argument would no longer bound the function to the sides of the player's virtual mask. Additionally, the player's collision functions can now detect collision with tilemap elements. As a result of all this, to re-enforce accuracy, as well as maintain similarity with the naming scheme of GameMaker's built-in mask intersection functions, the player's collision functions have been renamed as such:
* `player_in_object` -> `player_collision`
* `player_arms_in_object` -> `player_ray_collision`
* `player_body_in_object` -> `player_part_collision`
* `player_leg_in_object` -> `player_beam_collision`
Their JSDoc comments have also been updated to match these changes.