Unify terrain collision detection
With the exception of ceiling detection, all occurrences of the player checking for collision against tilemaps and solid zone objects would see the initialization of a `total_solids` temporary array that would merge the two, in order to streamline the procedure.

Due to how prominent this has been, and in order to reduce the amount of memory allocations and workload on the garbage collector for initializating temporary arrays, the player's `tilemaps` and `solid_objects` arrays have been fully merged into a new `solid_entities` array.

This array's starting values are the tilemaps the player is eligible to collide with. To seamlessly handle delisting of solid zone objects, a static variable has been initialized in the `player_detect_entities` function that stores the array's initial length, which is what it gets resized to. Evaluation of semisolid tilemap collision now takes place prior to zone object collision detection, and solid zone objects are pushed into the `solid_entities` array.

In the previous commit, the angle calculation function was modified to utilize GameMaker's ability to check for collision against multiple objects, instances, and tilemaps, by passing them as an array to the built-in collision functions. Isolated benchmarking revealed that this proved to be faster than manually iterating through the array's elements and checking for collision against each one, even when multiple collision function calls are used. However, this only turned out to be ever so slightly faster in practice. Nevertheless, to condense the amount of code, this behavior is now implemented for all occurrences of terrain collision detection. Furthermore, the `player_find_wall` function was rewritten to utilize the built-in `array_find_index` function, as this is faster than manual iteration of an array.

Due to the above change, the cliff detection procedure in the entering phase of the player's standing state has seen a rewrite, which also fixes a bug where if collision was not found with the ground at all, the player would switch to their "back" teeter animation anyway.

The only outlier in the removal of manual collision iteration was the player's ceiling detection function, as it still needs to check for semisolids; nevertheless, with the unification of terrain collision detection, semisolid evaluation has been rewritten, and iteration through the `solid_entities` array now runs front-to-back instead of back-to-front; with the emphasis on using tilemaps for level design, this means the player will detect collision with them first, allowing the function to resolve faster.

As part of these changes, the test room has seen the reintroduction of the quarter-ellipse solid zone object for testing purposes.
Moreover, the `player_calc_ground_normal` function has been renamed to `player_calc_tile_normal`, and the description of its `rot` argument has been rewritten for better accuracy.
An additional author note was also written in the `player_detect_entities` function, highlighting a limitation of its semisolid tilemap evaluation.