Introduce alternate ground snapping method
Typically, the player checks for collision with floors while on the ground at a height of `y_radius + y_tile_reach`. This extra distance is necessary to allow them to stick to steeper slopes, such as ellipses. However, there is a flaw with this where the player will snap down to perpendicular slopes when moving off of shallow ones; e.g., if the player where to move off a flat surface connected to a quarter-pipe, they would snap to the quarter-pipe instead of falling off.

The Harmony Framework introduced a solution to this whereby, if the player is standing at a ramp edge, its normal is calculated, and if it is too steep, the player, from then on, will check for collision with floors at a height of `y_radius + 1`, which is the minimum distance needed to detect collision with shallow floors.

In Sonic the Hedgehog 2, 3, & Knuckles, the extra distance the player checks for collision with floors scales to their speed; the formula used is `min(4 + floor(abs(x_speed)), 14)` (courtesy of Sonic Retro). In most cases, the issue of snapping down to perpendicular slopes does not happen when the player is moving very fast; as such, the player now uses this formula, albeit slightly modified: 2 is used instead of 4, and the `y_tile_reach` variable is used instead of 14.

Harmony's solution for snapping down to perpendicular slopes has been kept, but commented out: if any issues arise with this new method, the old one will promptly be reintroduced.