Create tilemap lookup array
In preparation for the addition of zone objects which require the use of tilemaps, in order to cut down on the number of `layer_tilemap_get_id` calls, a `tilemaps` array has been set up in the zone controller object, which, as its name suggest, stores all the tilemaps for the "Collision" tile layers. If one of these layers does not exist, it is removed from the array.

The player object's `solid_entities` array is now initialized by copying the zone controller object's `tilemaps` array, and registration of the tilemaps for the "CollisionSemisolid" and "CollisionPlane1" layers has been rewritten accordingly by checkng the length of the array.