Modify `player_beam_collision` to return IDs
This introduces the following changes:
* All occurrences of the `player_beam_collision` function have had their boolean checks rewritten to check against `noone`.
* The `player_find_wall` function has been completely removed; wall collisions are now checked by calling `player_beam_collision(solid_entities)`.
* The `player_find_ceiling` function has been slightly rewritten.

The `player_beam_collision` function has been moved to the bottom of the `PlayerCollisionFunctions` script, to group the other functions by return type.
An author note has also been added to the `PlayerTerrainFunctions` script which documents the change in checking for wall collisions, as well as a simple way to check for the presence of a floor or ceiling, without needing to align to one.